+++ src/qcommon/q_shared.c
@@ -1907,6 +1907,57 @@ Q_strupr(qchar *s1)
   return s1;
 }
 
+/*
+  Description: Replace strip[x] in string with repl[x] or remove characters entirely
+  Mutates: string
+  Return: --
+  Examples: Q_strstrip("Bo\nb is h\rairy!!", "\n\r!", "123"); //"Bo1b is h2airy33"
+  Examples: Q_strstrip("Bo\nb is h\rairy!!", "\n\r!", "12"); //"Bo1b is h2airy"
+  Examples: Q_strstrip("Bo\nb is h\rairy!!", "\n\r!", NULL);	// "Bob is hairy"
+*/
+void
+Q_strstrip(qchar *string, const qchar *strip, const qchar *repl)
+{
+  qchar *out = string;
+  qchar *p = string;
+  qchar c;
+  const qchar *s = strip;
+  const qint replaceLen = repl ? strlen(repl):0;
+  qint offset = 0;
+  qbool recordChar = qtrue;
+
+  while((c = *p++) != '\0')
+  {
+    recordChar = qtrue;
+
+    for(s = strip;*s;s++)
+    {
+      offset = s - strip;
+
+      if (c == *s)
+      {
+        if (!repl || offset >= replaceLen)
+        {
+          recordChar = qfalse;
+        }
+        else
+        {
+          c = repl[offset];
+        }
+
+        break;
+      }
+    }
+
+    if (recordChar)
+    {
+      *out++ = c;
+    }
+  }
+
+  *out = '\0';
+}
+
 
 // never goes past bounds or leaves without a terminating 0
 void Q_strcat( qchar *dest, qint size, const qchar *src ) {