+++ src/qcommon/qcommon.h
@@ -565,9 +565,7 @@ Cmd_ArgsBuffer(qchar *buffer, qint bufferLength);
 qchar *
 Cmd_Cmd(void);
 void
-Cmd_Args_Sanitize(void);
-void
-Cmd_Args_Sanitize2(size_t length, const qchar *strip, const qchar *repl);
+Cmd_Args_Sanitize(const qchar *separators);
 // The functions that execute commands get their parameters with these
 // functions. Cmd_Argv () will return an empty string, not a NULL
 // if arg > argc, so string operations are allways safe.